Move preparePathForTerminal to terminalEnvironment
